!clean up installed headers
JIRA: RTOS-125

+++ hal/ia32/string.c
@@ -116,7 +116,7 @@ int hal_strcmp(const char *s1, const char *s2)
 }
 
 
-int hal_strncmp(const char *s1, const char *s2, unsigned int count)
+int hal_strncmp(const char *s1, const char *s2, size_t count)
 {
 	const unsigned char *us1 = (const unsigned char *)s1;
 	const unsigned char *us2 = (const unsigned char *)s2;
